[Gold III] Title: 조삼모사, Time: 148 ms, Memory: 17644 KB -BaekjoonHub

+++ 백준/Gold/1621. 조삼모사/조삼모사.cc
@@ -3,7 +3,7 @@
 
 using namespace std;  
 
-tuple<long long,int,int> dp[1000001]; // 값, 개수, 추적
+tuple<int,int,int> dp[1000001];
 int N, K, C, arr[1000001];
 
 int main(){
